Moved connections into their own packages to allow consistent functional options.
